refactor: change calling search method from course service to course resolver
